// src/domain/enums/ExpenseCategory.ts
export enum ExpenseCategory {
  FUEL = 'FUEL',
  MAINTENANCE = 'MAINTENANCE',
  OTHER = 'OTHER',
}

export enum MaintenanceType {
  PREVENTIVE = 'PREVENTIVE',
  CORRECTIVE = 'CORRECTIVE',
}

export enum OtherExpenseCategory {
  INSURANCE = 'INSURANCE',
  TAXES = 'TAXES',
  CAR_WASH = 'CAR_WASH',
  PARKING = 'PARKING',
  TOLL = 'TOLL',
  FINE = 'FINE',
  ACCESSORY = 'ACCESSORY',
  DOCUMENTATION = 'DOCUMENTATION',
  OTHER = 'OTHER',
}